# src/api/api.yaml
openapi: 3.0.0
paths:
...
  /user/create:
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      tags: &ref_3
        - User
      security: &ref_4
        - bearer: []
  /user/list:
    get:
      operationId: UserController_getList
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of users with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponseDto'
      tags: *ref_3
      security: *ref_4
  /user/me:
    patch:
      operationId: UserController_update
      parameters: []
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      tags: *ref_3
      security: *ref_4
  /user/settings:
    get:
      operationId: UserController_getSettings
      parameters: []
      responses:
        '200':
          description: Current user settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponseDto'
      tags: *ref_3
      security: *ref_4
    patch:
      operationId: UserController_updateSettings
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserSettingsDto'
      responses:
        '200':
          description: User settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponseDto'
      tags: *ref_3
      security: *ref_4
  /user/role:
    patch:
      operationId: UserController_updateAdmin
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserAdminDto'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      tags: *ref_3
      security: *ref_4
  /user/{id}:
    delete:
      operationId: UserController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User soft-deleted (account disabled, sessions revoked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
      tags: *ref_3
      security: *ref_4
...
  /session/my:
    get:
      operationId: SessionController_getAllForUser
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of all sessions for the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponseDto'
      tags: &ref_6
        - Session
      security: &ref_7
        - bearer: []
  /session/user/{userId}:
    get:
      operationId: SessionController_getAllForUserByAdmin
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of all sessions for a specific user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponseDto'
      tags: *ref_6
      security: *ref_7
  /session/all:
    get:
      operationId: SessionController_getAll
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of all sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponseDto'
      tags: *ref_6
      security: *ref_7
  /session/current:
    get:
      operationId: SessionController_getCurrent
      parameters: []
      responses:
        '200':
          description: Get current session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponseDto'
      tags: *ref_6
      security: *ref_7
  /session/others:
    delete:
      operationId: SessionController_revokeOthers
      parameters: []
      responses:
        '200':
          description: Revoke all of the current user sessions except this one
      tags: *ref_6
      security: *ref_7
  /session/admin/{id}:
    delete:
      operationId: SessionController_revokeByAdmin
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Revoke any session (admin)
      tags: *ref_6
      security: *ref_7
  /session/{id}:
    delete:
      operationId: SessionController_revokeOwn
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Revoke one of your own sessions
      tags: *ref_6
      security: *ref_7
...
info:
  title: WHCP Backend
  description: XD
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
    JWT-refresh:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT refresh
      description: Enter JWT refresh token
      in: header
    JWT-reset-password:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT reset password
      description: Enter JWT reset password token
      in: header
    JWT-register-user:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT register user
      description: Enter JWT register user
      in: header
  schemas:
...
    PaginationDto:
      type: object
      properties:
        take:
          type: number
          minimum: 1
          maximum: 20
        skip:
          type: number
          minimum: 0
          default: 0
...
    Role:
      type: string
      enum:
        - USER
        - MODERATOR
        - ADMIN
        - OWNER
    UserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
      required:
        - email
        - role
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    UserListResponseDto:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - users
        - total
    UserSettingsResponseDto:
      type: object
      properties:
        id:
          type: string
        serverStatusEmailNotifications:
          type: boolean
        serverIdleEmailNotifications:
          type: boolean
        serverPushNotifications:
          type: boolean
        processEmailNotifications:
          type: boolean
        processPushNotifications:
          type: boolean
      required:
        - id
        - serverStatusEmailNotifications
        - serverIdleEmailNotifications
        - serverPushNotifications
        - processEmailNotifications
        - processPushNotifications
    PatchUserSettingsDto:
      type: object
      properties:
        serverStatusEmailNotifications:
          type: boolean
        serverIdleEmailNotifications:
          type: boolean
        serverPushNotifications:
          type: boolean
        processEmailNotifications:
          type: boolean
        processPushNotifications:
          type: boolean
    PatchUserAdminDto:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
...
    SessionResponseDto:
      type: object
      properties:
        id:
          type: string
        browser:
          type: string
          nullable: true
        os:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        isCurrent:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - updatedAt
    SessionListResponseDto:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - sessions
        - total
...